1/25 id Long으로 변경, 공지에서도 files 가져오기

+++ src/main/java/com/example/lollandback/gameBoard/mapper/BoardMapper.java
@@ -117,13 +117,13 @@ SELECT *,
              LEFT JOIN lolland.gameboardcomment gc on gb.id = gc.game_board_id
 WHERE gb.id=#{id}
 """)
-    GameBoard selectById(Integer id);
+    GameBoard selectById(Long id);
 
     @Delete("""
          DELETE FROM gameboard
          WHERE id=#{id}
             """)
-    int deleteById(Integer bgId);
+    int deleteById(Long bgId);
 
     @Update("""
 UPDATE gameboard
@@ -133,7 +133,7 @@ WHERE id= #{id}
     int update(GameBoard gameBoard);
 
     @Update("UPDATE gameboard SET board_count = board_count + 0.5 WHERE id = #{id}")
-    void boardCount(Integer id);
+    void boardCount(Long id);
 
     @Select("""
             <script>
